refactor: quickOpen extensions into integrated extensions folder
