+++ src/extensions/default/DefaultExtensions.json
@@ -21,9 +21,6 @@
     "NavigationAndHistory",
     "Phoenix-prettier",
     "PrefsCodeHints",
-    "QuickOpenCSS",
-    "QuickOpenHTML",
-    "QuickOpenJavaScript",
     "QuickView",
     "RecentProjects",
     "SVGCodeHints",
