--- src/extensions/default/QuickOpenCSS/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "QuickOpenCSS",
-  "version": "1.0.0",
-  "dependencies": {
-  }
-}
