--- src/extensions/default/QuickOpenCSS/requirejs-config.json
@@ -1,1 +1,0 @@
-{}
