--- src/extensions/default/QuickOpenHTML/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "QuickOpenHTML",
-  "version": "1.0.0",
-  "dependencies": {
-  }
-}
