--- src/extensions/default/QuickOpenHTML/requirejs-config.json
@@ -1,1 +1,0 @@
-{}
