--- src/extensions/default/QuickOpenJavaScript/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "QuickOpenJavaScript",
-  "version": "1.0.0",
-  "dependencies": {
-  }
-}
