--- src/extensions/default/QuickOpenJavaScript/requirejs-config.json
@@ -1,1 +1,0 @@
-{}
