+++ src/extensionsIntegrated/QuickOpen/html.js
@@ -19,15 +19,15 @@
  *
  */
 
+/*jslint regexp: true */
+
 define(function (require, exports, module) {
 
 
-    var EditorManager       = brackets.getModule("editor/EditorManager"),
-        QuickOpen           = brackets.getModule("search/QuickOpen"),
-        QuickOpenHelper     = brackets.getModule("search/QuickOpenHelper"),
-        JSUtils             = brackets.getModule("language/JSUtils"),
-        DocumentManager     = brackets.getModule("document/DocumentManager"),
-        StringMatch         = brackets.getModule("utils/StringMatch");
+    const QuickOpen           = require("search/QuickOpen"),
+        QuickOpenHelper     = require("search/QuickOpenHelper"),
+        DocumentManager     = require("document/DocumentManager"),
+        StringMatch         = require("utils/StringMatch");
 
 
    /**
@@ -37,55 +37,65 @@ define(function (require, exports, module) {
     * @param {number} line
     * @param {number} chFrom column start position
     * @param {number} chTo column end position
-    * @param {string} functionName
+    * @param {string} id
     */
-    function FileLocation(fullPath, line, chFrom, chTo, functionName) {
+    function FileLocation(fullPath, line, chFrom, chTo, id) {
         this.fullPath = fullPath;
         this.line = line;
         this.chFrom = chFrom;
         this.chTo = chTo;
-        this.functionName = functionName;
+        this.id = id;
     }
 
     /**
-     * Contains a list of information about functions for a single document.
-     *
-     * @return {?Array.<FileLocation>}
+     * Returns a list of information about ID's for a single document. This array is populated
+     * by createIDList()
+     * @type {?Array.<FileLocation>}
      */
-    function createFunctionList() {
+    function createIDList() {
         var doc = DocumentManager.getCurrentDocument();
         if (!doc) {
             return;
         }
 
-        var functionList = [];
+        var idList = [];
         var docText = doc.getText();
         var lines = docText.split("\n");
-        var functions = JSUtils.findAllMatchingFunctionsInText(docText, "*");
-        functions.forEach(function (funcEntry) {
-            functionList.push(new FileLocation(null, funcEntry.nameLineStart, funcEntry.columnStart, funcEntry.columnEnd, funcEntry.label || funcEntry.name));
-        });
-        return functionList;
-    }
 
+        var regex = new RegExp(/\s+id\s*?=\s*?["'](.*?)["']/gi);
+        var id, chFrom, chTo, i, line;
+        for (i = 0; i < lines.length; i++) {
+            line = lines[i];
+            var info;
+            while ((info = regex.exec(line)) !== null) {
+                id = info[1];
+                // TODO: this doesn't handle id's that share the
+                // same portion of a name on the same line or when
+                // the id and value are on different lines
+                chFrom = line.indexOf(id);
+                chTo = chFrom + id.length;
+                idList.push(new FileLocation(null, i, chFrom, chTo, id));
+            }
+        }
+        return idList;
+    }
 
 
     /**
      * @param {string} query what the user is searching for
-     * @param {StringMatch.StringMatcher} matcher object that caches search-in-progress data
      * @return {Array.<SearchResult>} sorted and filtered results that match the query
      */
     function search(query, matcher) {
-        var functionList = matcher.functionList;
-        if (!functionList) {
-            functionList = createFunctionList();
-            matcher.functionList = functionList;
+        var idList = matcher.idList;
+        if (!idList) {
+            idList = createIDList();
+            matcher.idList = idList;
         }
         query = query.slice(query.indexOf("@") + 1, query.length);
 
         // Filter and rank how good each match is
-        var filteredList = $.map(functionList, function (fileLocation) {
-            var searchResult = matcher.match(fileLocation.functionName, query);
+        var filteredList = $.map(idList, function (fileLocation) {
+            var searchResult = matcher.match(fileLocation.id, query);
             if (searchResult) {
                 searchResult.fileLocation = fileLocation;
             }
@@ -100,8 +110,8 @@ define(function (require, exports, module) {
 
     QuickOpen.addQuickOpenPlugin(
         {
-            name: "JavaScript functions",
-            languageIds: ["javascript"],
+            name: "html ids",
+            languageIds: ["html"],
             search: search,
             match: QuickOpenHelper.match,
             itemFocus: QuickOpenHelper.itemFocus,
