+++ src/extensionsIntegrated/QuickOpen/javascript.js
@@ -19,16 +19,14 @@
  *
  */
 
-/*jslint regexp: true */
-
 define(function (require, exports, module) {
 
 
-    var EditorManager       = brackets.getModule("editor/EditorManager"),
-        QuickOpen           = brackets.getModule("search/QuickOpen"),
-        QuickOpenHelper     = brackets.getModule("search/QuickOpenHelper"),
-        DocumentManager     = brackets.getModule("document/DocumentManager"),
-        StringMatch         = brackets.getModule("utils/StringMatch");
+    const QuickOpen           = require("search/QuickOpen"),
+        QuickOpenHelper     = require("search/QuickOpenHelper"),
+        JSUtils             = require("language/JSUtils"),
+        DocumentManager     = require("document/DocumentManager"),
+        StringMatch         = require("utils/StringMatch");
 
 
    /**
@@ -38,65 +36,54 @@ define(function (require, exports, module) {
     * @param {number} line
     * @param {number} chFrom column start position
     * @param {number} chTo column end position
-    * @param {string} id
+    * @param {string} functionName
     */
-    function FileLocation(fullPath, line, chFrom, chTo, id) {
+    function FileLocation(fullPath, line, chFrom, chTo, functionName) {
         this.fullPath = fullPath;
         this.line = line;
         this.chFrom = chFrom;
         this.chTo = chTo;
-        this.id = id;
+        this.functionName = functionName;
     }
 
     /**
-     * Returns a list of information about ID's for a single document. This array is populated
-     * by createIDList()
-     * @type {?Array.<FileLocation>}
+     * Contains a list of information about functions for a single document.
+     *
+     * @return {?Array.<FileLocation>}
      */
-    function createIDList() {
+    function createFunctionList() {
         var doc = DocumentManager.getCurrentDocument();
         if (!doc) {
             return;
         }
 
-        var idList = [];
+        var functionList = [];
         var docText = doc.getText();
-        var lines = docText.split("\n");
-
-        var regex = new RegExp(/\s+id\s*?=\s*?["'](.*?)["']/gi);
-        var id, chFrom, chTo, i, line;
-        for (i = 0; i < lines.length; i++) {
-            line = lines[i];
-            var info;
-            while ((info = regex.exec(line)) !== null) {
-                id = info[1];
-                // TODO: this doesn't handle id's that share the
-                // same portion of a name on the same line or when
-                // the id and value are on different lines
-                chFrom = line.indexOf(id);
-                chTo = chFrom + id.length;
-                idList.push(new FileLocation(null, i, chFrom, chTo, id));
-            }
-        }
-        return idList;
+        var functions = JSUtils.findAllMatchingFunctionsInText(docText, "*");
+        functions.forEach(function (funcEntry) {
+            functionList.push(new FileLocation(null, funcEntry.nameLineStart, funcEntry.columnStart, funcEntry.columnEnd, funcEntry.label || funcEntry.name));
+        });
+        return functionList;
     }
+
 
 
     /**
      * @param {string} query what the user is searching for
+     * @param {StringMatch.StringMatcher} matcher object that caches search-in-progress data
      * @return {Array.<SearchResult>} sorted and filtered results that match the query
      */
     function search(query, matcher) {
-        var idList = matcher.idList;
-        if (!idList) {
-            idList = createIDList();
-            matcher.idList = idList;
+        var functionList = matcher.functionList;
+        if (!functionList) {
+            functionList = createFunctionList();
+            matcher.functionList = functionList;
         }
         query = query.slice(query.indexOf("@") + 1, query.length);
 
         // Filter and rank how good each match is
-        var filteredList = $.map(idList, function (fileLocation) {
-            var searchResult = matcher.match(fileLocation.id, query);
+        var filteredList = $.map(functionList, function (fileLocation) {
+            var searchResult = matcher.match(fileLocation.functionName, query);
             if (searchResult) {
                 searchResult.fileLocation = fileLocation;
             }
@@ -111,8 +98,8 @@ define(function (require, exports, module) {
 
     QuickOpen.addQuickOpenPlugin(
         {
-            name: "html ids",
-            languageIds: ["html"],
+            name: "JavaScript functions",
+            languageIds: ["javascript"],
             search: search,
             match: QuickOpenHelper.match,
             itemFocus: QuickOpenHelper.itemFocus,
