+++ src/extensionsIntegrated/loader.js
@@ -33,6 +33,7 @@
 define(function (require, exports, module) {
     require("./icons/main");
     require("./RemoteFileAdapter/main");
+    require("./QuickOpen/main");
     require("./Phoenix/main");
     require("./InAppNotifications/main");
     require("./NoDistractions/main");
